Reset USB in HID mode if flashing is requested
Currently, the HID bootloader checks if the application has written
a magic word in the backup registers in order to reset the USB
connection in HID mode. However, if we loaded a faulty application
that does not write the backup registers, USB is not reinitialized
and flashing is no longer possible, even if we use the BOO1 button.

This change fixes the issue by resetting the USB to HID mode every
time we enter flashing mode, no matter the origin of the request.